+++ src/main/java/com/plutus/system/repository/AccountRepository.java
@@ -4,6 +4,8 @@ import com.plutus.system.model.entity.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Long> {
+public interface AccountRepository extends JpaRepository<Account, BigInteger> {
 }
